Split schemas and created seperate Thinky.js file
